# openapi_updated.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Social Trends API",
    "description": "API per aggregare trend da TikTok e Instagram",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/auth/generate-key": {
      "post": {
        "tags": [
          "🔑 Authentication"
        ],
        "summary": "Generate Api Key",
        "description": "🔑 **Genera nuova API Key**\n\nCrea una nuova API key per accedere all'API.\n\n**Piani disponibili:**\n- **Free**: 1.000 chiamate/mese, solo endpoint /global\n- **Developer**: 10.000 chiamate/mese, tutti gli endpoint base\n- **Business**: 50.000 chiamate/mese, include analytics avanzate\n- **Enterprise**: 200.000 chiamate/mese, supporto prioritario",
        "operationId": "generate_api_key_v1_auth_generate_key_post",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Email per associare la chiave",
              "title": "Email"
            },
            "description": "Email per associare la chiave"
          },
          {
            "name": "tier",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Tier del piano",
              "enum": [
                "free",
                "developer",
                "business",
                "enterprise"
              ],
              "default": "free",
              "title": "Tier"
            },
            "description": "Tier del piano"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/usage": {
      "get": {
        "tags": [
          "🔑 Authentication"
        ],
        "summary": "Get Usage Stats",
        "description": "📊 **Statistiche utilizzo**\n\nMostra le statistiche di utilizzo della tua API key.",
        "operationId": "get_usage_stats_v1_auth_usage_get",
        "parameters": [
          {
            "name": "api_key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "La tua API key",
              "title": "Api Key"
            },
            "description": "La tua API key"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Usage Stats V1 Auth Usage Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiKeyInfo": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "tier": {
            "type": "string",
            "title": "Tier"
          },
          "monthly_limit": {
            "type": "integer",
            "title": "Monthly Limit"
          },
          "current_usage": {
            "type": "integer",
            "title": "Current Usage"
          }
        },
        "type": "object",
        "required": [
          "key",
          "tier",
          "monthly_limit",
          "current_usage"
        ],
        "title": "ApiKeyInfo"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}